requirements-prod.txt
# ========================================
# ROBOTRADER - REQUIREMENTS PRODUÇÃO
# Ubuntu Server 22.04
# ========================================

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL
dj-database-url==2.1.0

# Cache
django-redis==5.4.0
redis==5.0.1

# Celery (Trading Bot)
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WSGI Server
gunicorn==21.2.0
gevent==23.9.1  # Async workers

# Exchanges (Trading)
ccxt==4.1.60

# Payments
stripe==7.8.0
mercadopago==2.2.1

# Security
cryptography==41.0.7
python-dotenv==1.0.0

# Utils
requests==2.31.0
Pillow==10.1.0
whitenoise==6.6.0  # Static files

# Streamlit (Dashboard)
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2

# Email
sendgrid==6.11.0

# Monitoring (Opcional)
sentry-sdk==1.38.0

# Performance
pytz==2023.3.post1
python-dateutil==2.8.2



